--- src/plugin.yml
@@ -1,8 +1,0 @@
-main: net.simplyrin.notgriefing.Main
-name: NotGriefing
-version: 1.1
-authors: [Sui(@Sui_pw),SimplyRin]
-commands:
-  notgriefing:
-    usage: /notgriefing <reload>
-    description: NotGriefing main command.
